requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
jinja2
requests
langchain-ollama          # Best integration for Ollama
openai                    # Install openai
aisuite[ollama]           # Optional fallback
# for all providers (recommended for flexibility):
# aisuite[all]
pandas
numpy
scikit-learn
shap
plotly
faker
streamlit
python-multipart